Support required nodes
Currently, the only way to specify that a given node is required for a
given `LaunchDescription` is to register an `OnProcessExit` handler for
the exit event of the required node that then emits a `Shutdown` event.
This use-case is pretty common; add some syntactic sugar to make it a
bit easier and reduce boilerplate.

Resolve #174

Signed-off-by: Kyle Fazzari <kyle@canonical.com>