Don't create a log directory every time the launch logger is imported.
To do this, we revamp the launch_config() function into a class
with member variables and methods.  We then make the 'log_dir'
a property, and if it is None when it is accessed, we create
the directory then.  This ensures that we will not create the
directory until someone actually wants to use it.

Signed-off-by: Chris Lalancette <clalancette@openrobotics.org>