[Dashing] Restore support for Python 3.5 (#324)
* Fix Python 3.5 compatibility by removing unsupported type hints.

Some of the type hints introduced in Dashing are not compatible with
Python 3.5. Where possible I tried to convert the hints to type comments
but for the type hints within function argument lists that did not
appear to be possible.

This is going to make backporting future changes more challenging since
we will have to:
1. Review the backported code for incompatible type hints
2. Reconcile merge conflicts caused by the removal of type hints on
the Dashing development branches.

* Convert list comprehension to generator (#300)

Addresses flake8 C412 errors introduced by flake8-comprehension 2.2.0

* Remove type comments longer than 100 chars.

It isn't clear to me from [PEP-0484] how to reconcile overlong lines
with type comments.

[PEP-0484]: https://www.python.org/dev/peps/pep-0484/

* Don't warn on imports used for type comments.

According to [PEP-0484] types used in type comments must be imported or
defined in the current module. Since flake8 is not type-comment aware
(at least in Dashing PR jobs) I've noqa'd the imports.

[PEP-0484]: https://www.python.org/dev/peps/pep-0484/

* Remove invalid trailing comma.

* Fix a test failure due to a change in how pytest represents exceptions.

This is a cherry-pick of 7b133347323d18522b986916571a5e3a82cc0790 from
https://github.com/ros2/launch/pull/277 to resolve a test failing on CI.

Signed-off-by: Steven! Ragnarök <steven@nuclearsandwich.com>