Add the ability to assert in stdout or stderr. (#378)
* Add the ability to assert in stdout or stderr.

Since we are switching the underlying logging to be stderr,
this is going to be important for tests.  This commit
refactors assertInStdout() to be assertInStream(), which
takes in the stream to be asserted on.  It then recreates
assertInStdout() on top of that, and adds in assertInStderr().

It also changes assertWaitFor() to wait on the stderr stream
by default.

Finally, add in a method called "assertInDefaultStream".

Once we backport this to Eloquent, this will allow users to
write tests that work in both Eloquent and newer releases.

Signed-off-by: Chris Lalancette <clalancette@openrobotics.org>