Refactor run_async loop to wait on queued events and futures at the same time
Fixes ros2/launch_ros#169

Otherwise, it's possible to get into a hung state where we wait for an event,
even though there are no more events. This is because the check for an "idle"
state evaluates to "True" as we wait for some futures to complete.

By waiting for futures and events concurrently, we can avoid this problem.
Further, we don't have to wait for an event if there's nothing in the queue.

Signed-off-by: Jacob Perron <jacob@openrobotics.org>