Add in pytest 7 compatibility.
In pytest 7, a bunch of APIs around hooks were deprecated.
Since we use those APIs, we are getting warnings every time
we invoke the APIs in question.  This commit adds in a compatibility
layer so that we can support pytest from 4.6.9 (which is in
Ubuntu 20.04), up to pytest 7.0.0 (which is what is currently
in pip).

Signed-off-by: Chris Lalancette <clalancette@openrobotics.org>