Addresses issue #588, allowing dict for 'output'
In issue #577 , the 'output' option for logging in launch
files was not participating in $(var) expansion (e.g., in
xml launch files).

The fix for this issue (commit 8006dfadc) appears to
have broken the use-case where the 'output' is specified
as a dictionary (e.g., in python launch files).

This commit checks if the output value is a dict
and only does the command expansion if it is not.

The 'output' value ultimately is passed to
launch/launch/logging/__init__.py: get_output_loggers()
which can accept either a string or a dict to specify
the outputs

Signed-off-by: Matthew Elwin <elwin@northwestern.edu>
(cherry picked from commit 7b27e2345f6c205f8f88974fd2666c658beb3390)