Fix the restoring of os.environ to maintain type. (#656)
* Fix the restoring of os.environ to maintain type.

During the LaunchContext constructor or _push_environment(),
we take a copy of the environment via os.environ.copy().  That
function returns a dict, but it turns out that os.environ itself
is an os._Environ object, not a dict.  Thus, when restoring the
environment in either _pop_environment() or _reset_environment(),
we cannot just set os.environ to that value.  Instead, we clear
out the environment and then set the key/value pairs one-by-one
into the environment.  That maintains os.environ as a os._Environ
object while also performing the desired pop or reset of the
environment variables.

Signed-off-by: Chris Lalancette <clalancette@openrobotics.org>
Co-authored-by: Jacob Perron <jacob@openrobotics.org>
(cherry picked from commit a36e9b945076ba9c0fbc1e9807600ed630a16ede)

# Conflicts:
#	launch/launch/launch_context.py
#	launch/test/launch/actions/test_push_and_pop_environment.py