[galactic] Support scoping environment variables (#601) (#631)
* Support scoping environment variables (#601)

* Support scoping environment variables

Resolves #597

Wherever actions were previously using os.environ, now get environment variables from the launch
context. There are a few exceptions where os.environ is still used since a launch context is not
available (e.g. logging directory and overriding launch process output).

Maintain a stack of environment variables in the launch context, similar to launch configurations.
Add new actions PushEnvironment and PopEnvironment to interact with the stack.
Push and pop the environment inside the GroupAction to support scoping.

Signed-off-by: Jacob Perron <jacob@openrobotics.org>

* Reset environment in GroupAction

Add a new ResetEnvironment action that sets the contexts environment
back to it's initial state (os.environ).

Signed-off-by: Jacob Perron <jacob@openrobotics.org>

* Operate directly on os.environ

Rather than keeping a copy of os.environ that entities must operate on, instead operate directly on os.environ.
This way if any Python interacting with the launch API also modifies os.environ, it should work as they expect.
Also keeps the behavior backwards compatible.

Signed-off-by: Jacob Perron <jacob@openrobotics.org>

* Fix the restoring of os.environ to maintain type. (#656)

* Fix the restoring of os.environ to maintain type.

During the LaunchContext constructor or _push_environment(),
we take a copy of the environment via os.environ.copy().  That
function returns a dict, but it turns out that os.environ itself
is an os._Environ object, not a dict.  Thus, when restoring the
environment in either _pop_environment() or _reset_environment(),
we cannot just set os.environ to that value.  Instead, we clear
out the environment and then set the key/value pairs one-by-one
into the environment.  That maintains os.environ as a os._Environ
object while also performing the desired pop or reset of the
environment variables.

Signed-off-by: Chris Lalancette <clalancette@openrobotics.org>
Co-authored-by: Jacob Perron <jacob@openrobotics.org>

* Sandbox environment in tests to fix repeated job failures (#609)

Signed-off-by: Shane Loretz <sloretz@osrfoundation.org>

* Fix bug in test_push_and_pop_environment.py

Sandbox the environment so the tests do not affect other tests.
This is a partial backport of https://github.com/ros2/launch/pull/652

Signed-off-by: Jacob Perron <jacob@openrobotics.org>

Signed-off-by: Jacob Perron <jacob@openrobotics.org>
Signed-off-by: Chris Lalancette <clalancette@openrobotics.org>
Signed-off-by: Shane Loretz <sloretz@osrfoundation.org>
Co-authored-by: Chris Lalancette <clalancette@openrobotics.org>
Co-authored-by: Shane Loretz <sloretz@osrfoundation.org>