Add default Python signal handler for all signals registered.
By default, and only in certain circumstances, SIGINT is the only signal
with a handler installed. All others will perform their function defined
by SIG_DFL, which is to ignore is some cases, and to terminate in others. In
either case, Python will never see them and be unable to pass them to
the async handler.

Signed-off-by: <cian.donovan@akara.ai>
Signed-off-by: Cian Donovan <cian.donovan@akara.ai>