Remove signal handler for SIGQUIT.
SIGQUIT signals are often generated by the user hitting 'CTRL+\' in an
interactive shell. However, when the terminal driver receives those
characters, it doesn't just send a SIGQUIT to the running process, but
to the entire foreground process group.

By default, the ROS2 Client Libraries don't register a handler for
SIGQUIT, as they do for SIGINT and SIGTERM, so the ROS2 node will be
terminated immediately and its core dumpted on receipt. So there is
little reason for ROS2 Launch to catch this signal for the purposes of forwarding it on to its children.

Also, not catching SIGQUIT means that if there is a problem with the
Python signal handling, then the user still has available a shortcut on
their keyboards to unilaterally kill the process.

Signed-off-by: <cian.donovan@akara.ai>
Signed-off-by: Cian Donovan <cian.donovan@akara.ai>