Replace individual handler functions with single generic one.
Allows arbitrary signal handlers to easily be added in future without
writing new functions, DRY.

Also allows at least one signal of each type, printing a warning when
more than one is received. This is to allow in future each signal type
to be passed to the children of launch, so a user can manually escalate
the kill signals.

Signed-off-by: Cian Donovan <cian.donovan@akara.ai>