Update the launch code for newer flake8 and mypy.
Newer flake8 complains if you use parentheses around
assert statements, so fix that.

Newer mypy had a number of minor complaints about our
types, so improve the type checking to account for that.

With both of these in place, the code is now clean under
the versions of flake8 and mypy in Fedora 38 (5.0.3 and 1.4.0,
respectively).

Signed-off-by: Chris Lalancette <clalancette@gmail.com>