Add in a timeout for launch pytests. (#725)
We recently had a PR where the tests for launch hung,
and this hung the CI job forever.  Avoid this by adding
in a 900 second timeout for the launch tests.  Given
that our current launch tests take ~200 seconds on our
slowest platform, this should be plenty of overhead
to ensure we have no "flakes".

Signed-off-by: Chris Lalancette <clalancette@gmail.com>