Rework task exceptions loop. (#755)
The original reason to do this was because mypy was
having a difficult time ensuring that the final raised
exception was valid.  This fixes that problem.

But this also should be much faster.  Before we were iterating
over the list 3 times; once to the collect the exceptions,
once to filter out None, and once to them print them out.
Instead, this just iterates once, dispatching everything
properly.

Signed-off-by: Chris Lalancette <clalancette@gmail.com>