Set due_to_sigint correctly
send_sigint? True : False
+----------------+--------+---------+
|                | SIGINT | SIGTERM |
+----------------+--------+---------+
| interactive    | False  | True    |
| noninteractive | True   | True    |
+----------------+--------+---------+

Launch either detects, through checking if there's a controlling terminal,
or the value of the flag `--noninteractive`, to determine if launch was
launched interactively or not.

If it was launched interactively, then the receipt of SIGINT
is implicitly assumed to have come from the user pressing Ctrl-C
and their terminal sending a SIGINT to the entire foreground process
group. In such case, it is redundant to also send SIGINT from launch to
the child ROS nodes, however it should be noted that even if another
signal was sent there would be no ill-effects.

In all other cases, assume the child nodes were not themselves
signalled, so manually send SIGINT, and later escalate to SIGTERM after
timeout, to the nodes.

Signed-off-by: Cian Donovan <cian.donovan@akara.ai>