Stop loading extensions during launch_{xml,yaml} tests.
The reason for this requires some explanation.  In current
Windows builds, a "soft" error is generated every time
these tests are run.  That error suggests that launch_xml or
launch_yaml failed to load in the launch_ros component.
But launch_xml and launch_yaml are in a repository that
launch_ros depends on; how are they trying to load in launch_ros?

The answer has to do with the extension mechanism.  Because
launch has the ability to load in extensions, it will attempt
to do so anytime the Parser class is initialized.  One of the
available extensions is the launch_ros parser, so it will attempt
to load that.  And on Windows, that will fail to find rclpy.  This
makes some sense; launch_xml doesn't depend on rclpy, so it doesn't
necessarily have the paths setup properly to find it.  However,
I don't understand why this only fails on Windows and not Linux.

Regardless, the solution presented here disables the loading of
extensions in both launch_xml and launch_yaml.  This should ensure
that it only loads in the things from this repository, avoiding
the problem completely.

Signed-off-by: Chris Lalancette <clalancette@gmail.com>